몬스터 이동까진 하는데 rotation을 ?

+++ GameServer/GameServer/GameGameThread.h
@@ -25,7 +25,7 @@ private:
 private:
 	//∏ÛΩ∫≈Õ
 	std::vector<Monster*> _monsters;
-	virtual void GameRun(int deltaTime) override;
+	virtual void GameRun(float deltaTime) override;
 	void SpawnMonster();
 	CObjectPool<Monster, false> _monsterPool;
 	int32 _maxMonsterNum = 5;
